pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.version]
path = "src/__init__.py"

[project]
name = "PHaSEAPI"
description = """
"""
authors = [
    { name = "Ludovico Boratto", email = "ludovico.boratto@acm.org" },
    { name = "Mirko Marras", email = "mirko.marras@acm.org" },
    { name = "Giacomo Medda", email = "jackm.medda@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dynamic = ["version"]
keywords = [
  "food recommendation",
  "api",
  "language models",
  "path modeling",
  "explainability"
]
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only"
]
dependencies = [
    "fastapi",
    "pydantic",
    "uvicorn",
    "polars",
    # "hopwise"
    "sentence-transformers>=4.1.0",
    "faiss-gpu>=1.7.2",
    "numpy==1.26.4",
    "hydra-core>=1.3.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "yellowbrick>=1.5",
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
  # Pylint
  "PL",
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # isort
  "I",
  # pyupgrade
  "UP",
  # pydocstyle
  # "D",
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"
